deprecate actors module, move submodules to root
