+++ src/tests/process_admin_cmds.rs
@@ -1,6 +1,6 @@
 use super::*;
-use crate::actors::api::{JsonResult, ResponseAccountState};
 use crate::adapters::admin::{process_admin, Command, RawFieldName, Response};
+use crate::api::{JsonResult, ResponseAccountState};
 use crate::primitives::{
     IdentityContext, IdentityFieldValue, JudgementStateBlanked, NotificationMessage,
 };
